Monitor Stream
Created a monitor class that will monitor the input bitrate.  If the bitrate goes to zero, it indicates the stream has stop.  The monitor class will wait based on the duration.   Once it reaches this timeout, the monitor class will close the input file.   This does not close the application process and is ready accept a new stream after the timeout.