Interupped Clipping (#6)
* Clipper Thread

- Created a Clipper class that will be run in a thread.  Moved all the code that was in main into this class.
- This project is now in the ConfBind solution so had to change the library paths.
- Avoid using Windows type names to C++ type names

* Stopping

Ensure files are closed when applications exits on Ctrl+C.

* Monitor Stream

Created a monitor class that will monitor the input bitrate.  If the bitrate goes to zero, it indicates the stream has stop.  The monitor class will wait based on the duration.   Once it reaches this timeout, the monitor class will close the input file.   This does not close the application process and is ready accept a new stream after the timeout.