17 handle looping (#18)
* Handle HEVC

Handle clipping HEVC encoded video.

* onCreateClip

An event handler called when a clip needs to created.

* Interim Commit

Create a segment of Access Units.

* Increase the Interval

Increase the interval from 900,000 to 27,000,000 to ensure the clip is created.

* Change Test Case

Use a more representative sample for the test case.
Remove the old sample transport file.

* Bump Version

Bump version to 1.0.3.

* Handle Looping

Detect when the PCR time loops.  If so close the file and open a new file.

* Bump Version

Bump version to 1.1.0 because of new command line option.  The changes are backward compatible.

* Key Frame

New command line option -k for key frame.  When present, the Mp2tClip instance will try to ensure the first frame in the clip is a key frame.

* Duration

If the duration for a clip is create than the input stream, create a clip file when closing.

* Ensure File is Open

Ensure the first file is open when creating a clip.

* Create 1st Clip File

Create 1st clip file when you receive the first TS packet.

* Boost::interprocess

Add Boost::interprocess file_lock class to lock the clip file.

* App Mode

Set the output clip file in append mode.

* Remove App Mode

* Improve Writing

If key frame option is disable, just write out to the clip file.
Else if enable, write out to the clip file until you have a key frame then save it to a segment. The segment will be used to beginning the next clip file.

* Fix GCC compiler error

Fix compiler error with GCC 13.

* Update README.md

Update generate build env.
Update usage.